test: extract model-edge-cases uniques, delete duplicate coverage file
`tests/unit/model-edge-cases.test.ts` was 37 tests written "to improve
code coverage" but ~65% (24 tests) duplicated dedicated tests in
`investment-split.test.ts` and `item.test.ts`. A per-test audit found:

- 20 pure duplicates (identical branch coverage in the peer file)
- 3 FLAKY duplicates using `new Date()` at exact 30/31/0-day
  `isConsentExpiringSoon` boundaries — `item.test.ts` already covers
  the same branches with safe 10-day / 60-day offsets
- 4 genuinely-unique branch tests (split factor precedence, item
  status fallbacks) → migrated to peer files
- 9 unique Category-module tests → no peer file exists, migrated to
  a new `tests/unit/category.test.ts`

### Migrated

**New `tests/unit/category.test.ts`** (9 tests):
- `getCategoryDisplayName`: name/missing/undefined/empty-string/all-fields
- `CategorySchema`: minimal/optional-fields/rejects-missing-id/passthrough

**`tests/unit/investment-split.test.ts:getSplitMultiplier`** (+2 tests):
- factors-beat-split_ratio precedence (both present)
- split_ratio fallback when only one factor is present (incomplete pair)

**`tests/unit/item.test.ts:getItemStatusDescription`** (+2 tests):
- "Connection error" fallback when status='error' but error_message missing
- raw connection_status passthrough for unknown statuses like 'pending'

### Dropped (no migration)

- 20 pure duplicates — branch coverage is preserved by the peer file
- 3 flaky boundary tests — peer file already covers these branches
  with non-flaky inputs
- 2 redundant "minimal item" tests — covered by `item.test.ts:132 + :255`

### Net

- -426 lines, -24 tests (37 deleted, 13 migrated elsewhere)
- Zero unique behavior lost; boundary flakiness eliminated
- Category module now has a properly-named dedicated test file
